% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_variable.R
\name{aggregate_variable}
\alias{aggregate_variable}
\title{Aggregate an SCJS Variable}
\usage{
aggregate_variable(variable)
}
\arguments{
\item{variable}{}
}
\value{
A modified column in a data frame.
}
\description{
Aggregate an SCJS Variable
}
