% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_select_helper.R
\name{column_select_helper}
\alias{column_select_helper}
\title{Easily find all columns to load from SCJS data}
\usage{
column_select_helper(
  dataset_type,
  variable_list = NULL,
  sections = NULL,
  years_to_load = NULL
)
}
\arguments{
\item{dataset_type}{A short string specifying what type of SCJS dataset to look for columns in. Must be one of c("nvf").}

\item{variable_list}{A single variable or vector of variables of 'pipeline' names from the 'var_name' column in the variable map.}

\item{sections}{A single string or vector of a section in the 'section_or_module' column in the variable map. Used to import entire sections.}

\item{years_to_load}{A single year or range of years to look for the original variables in.}
}
\description{
Creates a list containing the original column names for all sections or harmonised variables you want to load from the raw data.
}
