% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs-package.R
\docType{package}
\name{scjs-package}
\alias{scjs}
\alias{scjs-package}
\title{scjs: Import, Harmonise and Analyse Scottish Crime and Justice Survey Data}
\description{
A package providing functions to import, harmonise and analyse data from the Scottish Crime and Justice Survey (SCJS). Integrates with data available from the UK Data Service (UKDS).
}
\author{
\strong{Maintainer}: Stuart Napier \email{stuart.napier@gov.scot}

Authors:
\itemize{
  \item Scottish Crime and Justice Survey Team \email{scjs@gov.scot}
}

Other contributors:
\itemize{
  \item Scottish Government \email{statistics.enquiries@gov.scot} [copyright holder, funder]
}

}
\keyword{internal}
