% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scjs_design_factor}
\alias{scjs_design_factor}
\title{SCJS Design Factor Information}
\format{
\subsection{\code{scjs_design_factor}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{year}{Numeric year when most of the fieldwork occurred}
\item{survey_year}{Formatted value for fieldwork year}
\item{design_factor}{Generic design factor associated with that survey}
...
}
}
}
\usage{
scjs_design_factor
}
\description{
A small data set containing the values of the generic design factor for each
single SCJS year
}
\keyword{datasets}
