% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs_load_ukds.R
\name{scjs_load_ukds}
\alias{scjs_load_ukds}
\title{Import data downloaded from the UKDS}
\usage{
scjs_load_ukds(
  path,
  dataset_type = "nvf",
  years_to_load = NULL,
  columns = NULL,
  name_prefix = NULL
)
}
\arguments{
\item{path}{A file path to where files downloaded from the UKDS are.}

\item{dataset_type}{A short string specifying what type of SCJS dataset to load. Must be one of c("nvf").}

\item{years_to_load}{An option to specify what years to try and load in. Must be numeric, can be a single year, vector or range.}

\item{columns}{Columns to specify to load in from the call to haven::read_sav()}

\item{name_prefix}{If you want to specify a name convention for the loaded data, defaults to dataset type plus year e.g. 'nvf_2024'.}
}
\value{
A string
}
\description{
Import data downloaded from the UKDS
}
