% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukds_lookup}
\alias{ukds_lookup}
\title{UKDS SCJS Name Lookup Table}
\format{
\subsection{\code{ukds_lookup}}{

A data frame with 12 rows and 3 columns:
\describe{
\item{year}{Numeric year when most of the fieldwork occurred}
\item{survey_year}{Formatted value for fieldwork year}
\item{nvf_folder}{Folder containing non-victim form (NVF) files for a given year}
\item{nvf_file}{Name of the SPSS file containing non-victim form (NVF data for a given year)}
...
}
}
}
\usage{
ukds_lookup
}
\description{
A lookup table containing the names of folders and files as they are when
downloaded from the UKDS.
}
\keyword{datasets}
