% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vm_nvf_overview}
\alias{vm_nvf_overview}
\title{SCJS NVF Variable Map Overview}
\format{
\subsection{\code{vm_nvf_overview}}{

A data frame with 463 rows and 21 columns:
\describe{
\item{section_or_module}{Name of the section where the sheet containing the exact recoding is given}
\item{var_name}{Name of the variable as it appears in the original SG SCJS data pipeline code}
...
}
}
}
\usage{
vm_nvf_overview
}
\description{
An overview of how variables have changed over time in SCJS data.
}
\keyword{datasets}
