% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pooled_weight.R
\name{generate_pooled_weight}
\alias{generate_pooled_weight}
\title{Create a table from SCJS data}
\usage{
generate_pooled_weight(
  dataset,
  weighting_var,
  output_suffix = "_pooled",
  keep_all_columns = FALSE
)
}
\arguments{
\item{dataset}{Input dataset to add the pooled weights to.}

\item{weighting_var}{Variable to base the pooled weights on (nominally the
name of the variable containing gross individual or household weights).}

\item{output_suffix}{The suffix of the name of the new pooled weight
variable. This is '\verb{_pooled}' by default.}

\item{keep_all_columns}{Boolean variable, \code{FALSE} by default, specifying
whether to keep the columns used to calculate the pooled weights if \code{TRUE}, or if
\code{FALSE} keeps just the resulting weight column.}
}
\value{
Returns a data frame object of the same dimensions of the input
dataset with an added column for the pooled weight, as well as additional
columns used for the calculation if \code{keep_all_columns=TRUE}.
}
\description{
Adds a pooled individual or household weight to a combined SCJS
dataset. Implements the same methodology outlined in Appendix B of the
published Scottish Government paper \href{https://www.gov.scot/publications/ethnicity-justice-system/pages/8/}{Ethnicity in the Justice System}.
}
