% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs_harmonise_variable.R
\name{scjs_harmonise_variable}
\alias{scjs_harmonise_variable}
\title{Harmonise variables from SCJS datasets}
\usage{
scjs_harmonise_variable(target_df)
}
\arguments{
\item{target_df}{The dataset that you want to harmonised variables to be added to}
}
\description{
Harmonises variables from individual SCJS datasets into a target single multi-year dataset.
}
