% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs_harmonise_variable.R
\name{scjs_harmonise_variable}
\alias{scjs_harmonise_variable}
\title{Harmonise variables from SCJS datasets}
\usage{
scjs_harmonise_variable(original_data, vars)
}
\arguments{
\item{original_data}{A list containing the original SCJS datasets to harmonise from}

\item{vars}{A vector of variables to be harmonised and placed into the target data frame}
}
\description{
Harmonises variables from individual SCJS datasets into a target single multi-year dataset.
}
