% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs_load_ukds.R
\name{scjs_load_ukds}
\alias{scjs_load_ukds}
\title{Import data downloaded from the UKDS}
\usage{
scjs_load_ukds(
  path,
  dataset_type = "nvf",
  years_to_load = NULL,
  columns = NULL,
  name_prefix = NULL
)
}
\arguments{
\item{path}{A file path to where files downloaded from the UKDS are.}
}
\value{
A string
}
\description{
Import data downloaded from the UKDS
}
\examples{
x <- "path/to/data"
scjs_load_ukds(x)
}
