% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scjs_table.R
\name{scjs_table}
\alias{scjs_table}
\title{Create a table from SCJS data}
\usage{
scjs_table()
}
\arguments{
\item{dataset}{Input dataset to construct the table from}

\item{table_type}{Specify what format the output table should take (time series / subgroup comparison etc.)}

\item{var}{Single variable or vector specifying multiple variables to be presented in the resulting table}
}
\description{
Takes an individual or harmonised set of SCJS data and builds a table using standard methods.
}
